// src/react/components/dashboard-visualization/components/TimeSeries2/types.ts
export type UseRefreshActionsParams = {
  resetZoom: () => void;
  setStoredChartState: (params: { newStoredChartState: Partial<StoredChartState> }) => void;
};

export type RefreshActionsParams = {
  key: string;
  chartData: ChartData;
  resetZoom: () => void;
  handlerSetters: ToolbarSetActionsHandlers<
    "DashboardCanvas-ViewModeNode-TimeSeries" | "DashboardCanvas-EditModeNode-TimeSeries"
  > | null;
};
// ...
export type LineChartData = {
  data: LineChartDataSeries;
  label: string;
  min: number;
  max: number;
  dataSetId: string;
};

export type RequestAndPrepareColumnData = (
  params: CommonParams<"TimeSeries"> & {
    dataRange: DataRange | null;
    pointCount: number;
  }
) => Promise<LineChartData[] | null>;

export type LineChartDataWithColor = LineChartData & { color: string };
// ...
export type HiddenColumnsKey = `${string}-${string}`;

export type HiddenColumnsMap = Record<HiddenColumnsKey, boolean>;

export type StoredChartState = {
  range: Range | null;
  showMarkers: boolean;
  hiddenColumns: HiddenColumnsMap;
  tooltipsEnabled: boolean;
  showMinimap: boolean;
};
// ...
export type ChartData = {
  resizeObserver: ResizeObserver | null;
  numtraces: number;
  xAxisLabel: string;
  yAxisLabel: string;
  range: Range | null;
  fullRange: Range;
  dataSetId: string | null;
  datasetLabels: string[];
  chartColumns: LineChartDataWithColor[];
  minimapChartColumns: LineChartDataWithColor[];
  chartSurface: SciChartSurface | null;
  wasmContext: TSciChart | null;
  overview: SciChartOverview | null;
  xySeriesArray: XyDataSeries[];
  minimapXySeriesArray: XyDataSeries[];
};